Move storage.rs
